Rewrite binary/perllib paths in final executables
The Makefile for MUMmer unhelpfully takes it's own build directory
and inserts that path into some of the final perl scripts. That is,
the final destination path has scripts that still try to load perl
library from the build path.

This fix identifies those files and lazily fixes the incorrect
paths. The fixes are 'lazy' because they use regexes that only
check the beginning part of the string, not the full string with
an erroneous build directory path. Ergo, a future update to the
code that has additional 'use lib [path]' directories might break.

That being said, the original authors will be contacted to see if
future releases can not assume that the build directory and the
install directory are the same.

Additionally, the Foundation.pm file needs to be copied out of the
build directory and into the install directory. Previously it wasn't.