added HMMER version 3.1b1
From the hmmer website:
"This code is expected to be stable, but we’re releasing it as a beta
test just to be careful. After some time in the wild, we’ll make a
release candidate, and if you folks haven’t chewed any of that up too
badly, we’ll make the final 3.1 release reasonably soon."

I added runtest="check", which unveiled a bug in the build procedure of
the unit test. The attached patch solves that (bug reported upstream).