Remove sanity check for libnpp for CUDA 5.5 compat
CUDA 5.5 splits libnpp into three libraries: libnppc, libnppi and
libnpps. This causes the sanity check for libnpp.so to fail on version
5.5.22.

Simply removed this check to maintain compatibility with CUDA 5.0. It's
debatable whether it's better to special-case this check by version, but
given that the CUDA installer is monolithic we probably don't need this
check in addition to the others.

Also removed executable bit from cuda.py permissions for consistency
with other EasyBlocks.