Fix OpenMPI OS dependencies for Debian and Ubuntu
OpenMPI-1.6.4-GCC-4.7.2.eb had an OS dependency on libibverbs-devel,
which is a package found in RHEL-based distros and Fedora. In Debian
and Ubuntu, this package is called libibverbs-dev. In this commit I
added an OS_NAME check to make the libibverbs-devel check conditional
on running RHEL, etc. and added a check for the correct package name
on Debian.

I also added the same check into OpenMPI-1.6.4-GCC-4.6.4.eb because
it has the same dependency, and this easyconfig is used for the
goolfc-1.3.12 toolchain.