Create texinfo 4.13a easyconfig
Create a texinfo 4.13 easyconfig based on the newer 5.2 easyconfig
to support builing older GCC (<=4.7.2) on newer systems. The 4.x releases
of texinfo tolerated a document parsing error in earlier versions of GCC.

In particular, loading the resulting module allows building GCC 4.7.2
on CentOS7.  This GCC is the foundation of the goolf-1.4.10 toolchain
used by many bioinformatics packages in the current EasyBuild (2.6).

The easyconfig's GCC requirement was set to 4.8.4 as a convinience to match
the version of GCC likely already built by the goolf-1.7.20 toolchain.