Removed buildininstall=true
it seemed to not do what I wanted. It created a subdirectory under the
install directory where the freebayes tarball was built and then
installed up one level. It also can not fix the problems with hardcoded
relative paths as this depends on where you run various parts of the
software from.