Update FSLeyes to 0.20.0 (w dep modules)
Update FSLeyes to the latest version.  Prior Bundle easyblock changed to
PythonPackage, with extensions changed to regular dependencies (some
explicit).

The rationale for reshuffling is that the setup.py for this version of FSLeyes
seems to be installing some of its dependencies into the build tree (instead
of the install tree), causing import failures at runtime.  This doesn't fix
the root problem, but works around it.

Also, set $FSLDIR to avoid app error dialog.