ignore missing binutils dependency for binaries
By default the testsuite checks for a binutils or GCC dependency, when
using the GCCcore toolchain to avoid accidentially using the system
binutils. However, this is not necessary for packages that do not need
to be build like binaries. This commit adds exceptions for the 'Binary'
and the 'PackedBinary' easyblock to the existing exception for the
'Tarball' easyblock.