fixed and improved Amber(Tools) pathscript patches
AmberTools 21 and newer contain the QUICK QM-engine, for which
more lines were added to the amber.sh and amber.csh pathscripts.

This commit backports the patch written for AmberTools-21
to Amber/AmberTools-20 and renames the patches with QUICK
to indicate their content.
An additional comment has been added to the patches and the
checksums have been updated accordingly.