Refactor `check_dep_vars`
Remove the C&P code for filtering a dep_var with a specific
version_suffix if exactly 1 match was found.
Replace by a list similar to `multiple_allowed_variants` and the code to
directly delete the key instead of rebuilding the dict.
Also assert only once with all found issues.

Allows easier extension.