Remove flake8 suppressions for invalid escape sequences
Those will become a SyntaxError in some Python version,
so that suppression to make CI pass isn't a good idea,
because that sequence IS invalid, no matter what flake8 says.
Solution is either to use raw strings or remove the escapes.
For OmegaFold it isn't required as an escaped space is used just to make
for nice indentation which isn't required at this place.