fix CI check for extension patches using `alt_location`
While `alt_location` can be used for patches in the main EasyConfig part
and in extensions the verification code handled them differently and
missed to take the `alt_location` into account for extension patches.

Factor our the checking code to a function that can be used for both
cases.
This also removes a redundant condition in one of the code instances
before verifying the checksums.