Create OpenFOAM v2412 file from v2406 file
Upstream provided a plugins tarball for the v2406, but not v2412
or earlier releases. These plugins were included in the v2406
easybuild though, so switch source to git checkout for v2412 in
order to continue building them.

Used the .txz format as easybuild said it was required for a stable
checksum. Rebuilding the source again produced the same checksum,
so it does appears to work.