{compiler}[GCCcore-13.2.0] nvidia-compilers-25.7, CUDA-12.4
Based on [#23125](https://github.com/easybuilders/easybuild-easyconfigs/pull/23125),
this tries out one possible variation of an easyconfig for the
nvidia-compilers. This is to make the module compatible with the
defaults in EESSI/2023.06. It swaps out the SYSTEM toolchain for
GCCcore-13.2.0 and uses that same gcccore version in the other deps.
Also default to CUDA-12.4.

Requires:
+ https://github.com/easybuilders/easybuild-easyblocks/pull/3788
+ https://github.com/easybuilders/easybuild-framework/pull/4927